+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -64,7 +64,7 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-02-12T11:56:41.781Z")
+    created: Optional[datetime] = parse_datetime("2026-02-12T13:43:57.250Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")