+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -186,7 +186,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh8vc382ef0rvf3cbwbhzkpe"
+        "conv_01kh91ggjaehq7cppnbejkbvqq"
     )
     r"""Unique conversation identifier with `conv_` prefix."""