+++ packages/orq-rc/RELEASES.md
@@ -5589,3 +5589,13 @@ Based on:
 - [python v4.3.0-rc.59] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.59] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.59 - packages/orq-rc
+
+## 2026-02-12 13:46:35
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.713.1 (2.821.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.60] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.60] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.60 - packages/orq-rc